Always export batch when limit reached (#519)
Currently spans will begin to drop if more than the max queue size are
generated within a given batch window. This change causes the batch to
export when full, allowing more spans to be exported within a given
window even if the scheduled export time has not yet been reached.